LabVIEW Delendum Est
FHP got the serial comms working properly including the ability to read large amount of data. Must write an interface for the IBM4 that can use serial comms only. Base the code on the existing VISA methods. Make the interface an object that calls methods, rather than a library that calls methods. This will be cleaner.