LabVIEW Delendum Est
Have implemented the ReadSingleChnnl, ReadMultipleChnnl and DiffRead methods. Reading is slow due to the requirement of having to clear the device buffer after every read operation. It's all working fine, but I wonder is the VISA comms that bit faster. It should be possible to check this. An alternative is to implement the class using VISA comms where you open and close the serial comms using the constructor. Is this actually faster? Testing required. Upgraded the gitignore file also.