LabVIEW Delendum Est
have implemented the various read methods in a single method similar to what has been done in LabVIEW. Must finish the implementation for the DiffRead method but otherwise all fine.

Timing of the measurements is not an issue like I thought it was. Variable sample of the IBM4 means that it's all fine. Proceed as you were.